request #12995: Move distributed SVN Tuleap end-to-end tests to the nightly test run
Due to the little development currently in this area it seems acceptable to play
this test suite only once a day.
This should however not be seen as a definitive thing, developers starting
a new development about the distributed Tuleap setup should not hesitate
to move back the tests in the pipeline played on each patchset.

Change-Id: I14a491e89a567f1b3b86f3f91dd7554a1ed39814