Use PSR-18 HTTP client interface instead of the interface provided by HTTPPlug
The end goal is to make easier to interact with the rest of the
PHP eco-system. PSR-18 [0] is still quite recent but it is likely that over
time libraries will rely on it as a first party component or at least
provide an adapter.

The usage of the ErrorPlugin has been dropped because it breaks the contract
sets by PSR-18.

No functionnal changes are expected.

This is part of request #12996: Move to PSR-18 HTTP client and PSR-17 HTTP factories

[0] https://www.php-fig.org/psr/psr-18/

Change-Id: I40b7b4cef5e9b8c59b6ce86a6a47d18553a26152