Bind nb max values and is project flag to key press
Go to siteadmin » project settings » categories. Edit a top level
category. On the nb max values input, hit arrows keys to switch from
value 3 to 1 and then from 1 to 3. => the Used as project flag checkbox
should be disabled when nb max ≠ 1.  The same if you use the mouse to
change the nb max value.

Bonus: the script was not compatible with IE11, this is fixed now.

Part of story #12553: see flags that categorize my project in all project pages

Change-Id: I62c6d3d670f2c878639527db3630b82e55dbb221