Use PSR-17 [0] HTTP factories
This replace the existing factories that had the same role.

With this contribution and the one introducing PSR-18 HTTP clients
(see d489d91ef0ab38a2610c1eab55a2351f9465453d) Tuleap has within its
internal API a modern way to interact with HTTP services. The use
of the PSR interfaces should be future proof and at least guarantee us
some interoperability with the rest of the ecosystem.

No changes are expected at the functional level.

This is part of request #12996: Move to PSR-18 HTTP client and PSR-17 HTTP factories

[0] https://www.php-fig.org/psr/psr-17/

Change-Id: Iaa70046a7559339f1ffec0e99a6247e66dbc1873