The new worklow UI is the default one
This is part of story #12187 configure workflow pre & post actions at target state level

All workflow administration views must use the new Vue.js app.

A flag can be defined to switch back to the old UI:
sys_tracker_whitelist_that_should_use_legacy_workflow_transitions_interface
with comma separated tracker ids.

If the workflow is switched to the old UI, the REST endpoints cannot be used
for them.

A warning message is displayed if a workflow is in simple mode and a
tracker administrator wants to access the legacy UI.

Change-Id: I58b640cca38c13ccb0499bd88a4f4000ea5a1510