request #13007: Some PHPUnit unit tests do not properly "close" Mockery
As the result the expection count per test can be incorrect and tests
can fail if they assert no expectation.

This contribution fixes the impacted tests and makes sure it can not
happen again. Tests using Mockery and not calling Mockery::close()
at the end will be marked as failed.

Change-Id: I6bc11ddb1735db2586abb4918aa7774c586204c9