Fix the display of quick look button and the quick look display
This is a part of the story #12450 see a preview of document

The display of the quick look should now stay on the screen when the user scroll
When the screen size is reduced the quick look is displayed under the item. It will be fixed later.

How to test:
  - Apply the patch
  - Click on the quick look of an item
 => The quick look should be displayed and if you scroll the quick look is always displayed.

Change-Id: I5d9eea96c0ed7cbb23280f4ca941a794639dd35b