Fix the error display when the back end fail
This contribution should fixes the front part of request #12997: Upload a large
file in new docman in error without error feedback

When there was a back end problem, in some case the error was not displayed.
For example: when a time-out occured, the UI dsiplays the empty folder screen.

How to to test:
 - Add a sleep(1000) to the REST route you want to test. ( i.e getDocumentsItem)
 - Go to the Document UI and refresh.
 - Wait ... or go take a cofee.
=> You should see the error screen and when you click on "see the details" the details appears

You can also test it in a modal.

Change-Id: I1ac62e27c8b66c79f3b2fa4eea5d93c53f4aa541