PATCH trackers/:id to switch to simple mode
This is part of story #12187 configure workflow pre & post actions at target state level

A workflow can now be switched from advanced to simple.
This implies that for each possible values in the workflow, we get the
first transition (best effort) and applies it to all other defined
transitions in the workflow.

The code is not well optimized. This first step is to make the feature
work. We will make it better in next commits.

Change-Id: I63fe90185350e0d547c0e041de4ba78b39e4b25e