request #13013: Internal Tuleap API should not expose multiple ways to work with DB transactions
There is currently multiple ways to achieve the same work with DB transactions.
This contribution tries to clean that so contributors are naturally guided
to the expected solution for the problem thay are trying to solve.

As a result, you can end up in two cases:
* you are working inside a \Tuleap\DB\DataAccessObject instance
and needs to do some work protected by a transaction. For example,
you want to delete/update rows according to some result you retrieve
locally. You are expected to use what's being proposed by the object
given by self->getDB().
* you are working elsewhere and have the need of rollback your DB updates
if something goes wrong. For example, you want to rollback if a file is not
written successfully. The use case might also to wrap some high level operations
to ensure the end result is consistent. You are expected to use an instance of
\Tuleap\DB\DBTransactionExecutor.

No functionnal change is expected from this contribution.

Change-Id: Ic1fbd43ebc99005e7dd9e7cd60f6012c37ee03e6