Fix the UI display when the upload finish fail
This contribution should fixe the front part of request #12997: Upload a large
file in new docman in error without error feedback

When an file fail the progress bar was always displayed and the fake item too.
This contribution should correct this case. Now the progress bar is no more displayed and the fake should disapear. A modal should also displayed when the upload fail.

How to test:
  - Go to DocumentDataStore.php and change the return of the "getFinisher()" function to null
  - Go to the new document UI and try to a uplaod a file. ( By DnD or via the modal )
=> At the end of the uploading you should see a modal which says that the upload failed.
=> The fake item should not be displayed at the end of the upload
=> The progress bar should be displayed anymore.

Change-Id: I4014182a393787fb4170653a25d86c02cb6dca74