request #13038: Introduce static analysis with Psalm
The overall goal of adding a static analysis tool is to
help contributors to write better code and gain confidence
when doing refactoring, especially the ones affecting a large
part of the codebase.

Psalm [0] is introduced to observe how it behaves with the contributions
currently integrated in Tuleap. The configuration is based on the "level 3" [1]
issues proposed by Psalm, it is still quite a permissive configuration but it
should catch most issues. The complete list of issues Psalm is capable to find
can be seen here [2].
The stage is only played in the nightly pipeline and errors will not mark the
pipeline has failed, it's only an observation phase for now.

Only new errors should be found by Psalm thanks to its baseline file, the
existing issues are marked with the level INFO.

Some files have been ignored from the analysis to not vaste resources on code
that should be removed and/or because Psalm crash when trying to analyze them.

A small wrapper to launch the tools has been developped so that contributors
with external plugins are not polluted by existing errors in those plugins.
External plugins should define their own configuration and (if needed)
generate their own baseline file [3] if they want to set up the analysis.

[0] https://github.com/vimeo/psalm/
[1] https://github.com/vimeo/psalm/blob/3.0.17/assets/config_levels/3.xml
[2] https://github.com/vimeo/psalm/blob/3.0.17/docs/issues.md
[3] https://github.com/vimeo/psalm/blob/3.0.17/docs/dealing_with_code_issues.md#using-a-baseline-file

Change-Id: Iec7edc618d5b260a131229741f9e1fd5a0997b84