request #13040 Remove & clean usage of PATH_INFO in Tuleap core
PATH_INFO is a part of standard HTTP request and was used as poor's people
beautiful urls when we didn't really understood how to configure an apache
server to do URL rewriting (good old days).

But we add to introduce quirks in nginx configuation to support that and
those quirks are now in conflict with front router deployment, especially
when we want to transparently replace old ".php" scripts by routes.

Change-Id: I39bf816f11a01725fa04ae2006f394a44da0304c