Creating two versions of a docman file within the same second via the REST API generates a crash
Docman_ItemDao::updateFromRow() tries to be smart and efficient by doing the
work of the DB. Elements of the row that has not been changed from the current
state in the DB will not be set in the SQL query. If no elements needs to be
updated the query will fail and consequently the file upload will fail too
(resulting in a failed REST test).

The code is neither smart nor efficient. It costs a SELECT query to maybe save
some column updates. The DB optimizer can decides of that on its own, we should
let it work.

This is part of story #12458: DnD on an existing file creates a new version

Change-Id: I2a2648b791ae95f300834565d04c0ecfcae6c7eb