request #13053: Setup code coverage for the PHPUnit unit tests test suite
The code coverage collection can be seen on the build results.

The code coverage whitelist in the PHPUNit configuration has been reworked
to include all the code and exclude only the problematic files.

PCOV [0] is used for the collect the code coverage. The configuration has not
been fine tuned, it seems to be good enough for now.
The pcov.exclude setting has however been tested with this PCRE:
'~^plugins/(?:\w|-)+/(?!.*?include).*/.*^src/(?!.*?(?:common|www)).*/.*^tests/(?!.*?lib).*/.*'
must it does not show a lot of improvements in our case and probably
no one wants to maintain it.

[0] https://github.com/krakjoe/pcov

Change-Id: I02387e28c3b7307199bb94fa1b411e442cd55155