request #13069 Move src/www scripts to front controller
src/www is a mess and it will takes time to move on an clean everything.
However we should not expose that by default to all new comers that enters
in the code base.

The LegacyRoutesCollector briges the 2 worlds by allowing declarationg of
old "script.php" routes as regular FrontRouter routes and redirect to the
old script. The old script is move in src/core so we can clearly mark this
place as "dangerous" and "deprecated".

Change-Id: I4fdfc39956f1806dcfe4d736342dec92ff47be4d