request #13068: Start enforcing Tuleap PHP coding standard in CI pipelines
Unused use statements are now rejected by the CI pipeline and the whole codebase
is fixed for this issue.

Two new helpers has been to the Makefile at the sources root to help
contributors launch phpcs and phpcbf.

To get a quick result, only the files modified by a contribution are tested.
The test pipeline will however check the whole codebase when played against the
master branch to help catching bad merges.

Change-Id: I10c36f9fd223eab4ded23ad3726fdcdeeefa3089