Allows calling REST route timetracking_reports/{id}/times without trackers or query
part of request #11800 Know the time spent on different trackers on a single view
little refactoring of REST route timetracking_reports/{id}/times to fix
request #13089 If all trackers are removed, some times continue to be here when no times expected

To reproduce it, remove all trackers selected from the widget Timetracking Overview and search.
On master, times are still displayed.
On this patch, trackers' times should be removed from the table.

If the route is calling with dates and without trackers' id : the route return times for report's trackers on given period.
If the route is calling without query : the route return times for report's trackers over the past month.

Change-Id: Id7ebaa7d8c549193811bd60c3df060031c466b80