request #13024: Deleting a reference generates errors
On projects where the size limitation for new objects is active,
deleting a reference in a Git repository generates warnings.

A patch has been proposed upstream but until the fix is released, Tuleap
should deploy its own VREF so no errors are displayed to users.

Note this change only come into effect once the Gitolite configuration
for repositories is regenerated.

Change-Id: I009dc97e36cb0e15760303b7b78b1870adf9702a