request #13093: Ensure DB connection is still up after archiving an artifact before its deletion
When an artifact is deleted, an archive is generated before the deletion really happen.
Archiving the artifact can take quite a long time without any DB interactions so the
DB server might decide to close the connection. If the connection is closed, the deletion
does not happen because the DB can not be reached anymore.

Change-Id: I1aef5bec82e65c216efce5ad0e451c688d60c804