Implements Tus protocol for FRS REST API
You can now use tus client to upload file in the FRS through REST API.

Implementation note
-------------------

The specificity of fileforge is that it depends on the utilitary script
fileforge.pl to be able to create files in folders owned by root
(fileforge has sticky bit).

However fileforge.pl need the real filename in the upload directory to
be able to move it. Therefore this patch is introducing the filename in
the TusFileInformation to allow FRS implementation to be able to create
file with real filename instead of just the id.

I'm not really proud of that, but I couldn't find a better way without
breaking the current implementation. Other options would be:
- update fileforge.pl to not rely on the filename in the upload
directory
- update access rights in /var/lib/tuleap/ftp/tuleap directory so that
codendiadm/tuleapadm can create files without having to rely on an
external script with sticky bit

This can be done in a dedicated changeset.

How to test
-----------
To test you can use the following snippet. You will have to provide the
path to the tus.js client library and update the release id to match
your environment. Upload information (progress) is displayed on the
devtools console. When the upload is done you can assert that the file
is present in the release through the web UI.

Note: after 1 second, the upload is automatically aborted (see timeout
at the end of the script), so that we can test the abort. In this case,
there isn't any leftovers in /var/tmp/frs/ongoin-upload dir.

```html
<input type="file" id="myfile">
<script src="tus.js"></script>
<script>
document.getElementById('myfile').addEventListener('change', async (e) => {

    const release_id = 1;

    const file = e.target.files[0];

    const post_response = await fetch('/api/v1/frs_files', {
        method: 'POST',
        credentials: 'same-origin',
        headers: {
            "Content-Type": "application/json"
        },
        body: JSON.stringify({
            release_id,
            name: file.name,
            file_size: file.size
        })
    });
    const post_response_body = await post_response.json();
    const upload_href = post_response_body["upload_href"];

    // Create a new tus upload
    const upload = new tus.Upload(file, {
        uploadUrl: upload_href,
        retryDelays: [0, 3000, 5000],
        metadata: {
            filename: file.name,
            filetype: file.type
        },
        onError: function(error) {
            console.log("Failed because: " + error);
        },
        onProgress: function(bytesUploaded, bytesTotal) {
            const percentage = (bytesUploaded / bytesTotal * 100).toFixed(2);
            console.log(bytesUploaded, bytesTotal, percentage + "%");
        },
        onSuccess: function() {
            console.log("Download %s from %s", upload.file.name, upload.url);
            alert('Done');
        }
    });

    // Start the upload
    upload.start();

    setTimeout(() => {
        upload.abort();
    }, 1000);
});
</script>
```

Part of story #12813: upload files with FRS REST API

Change-Id: I253b4190115548d2da727e64417a86c43b87054a