request #13098: Create test env plugin has missing dependencies to tracker and bot mattermost plugins
Bot Mattermost classes are not available when sending notifications resulting
in a crash if the administrator has set up at least one bot to send notification to.
Tracker plugin is not marked as an RPM dependency but files from the tracker
plugins are loaded, this can crash if the tracker plugin RPM is not installed.

Change-Id: I76a4f9a1442470c989c809e60f104432a35bb12a