Move Upload classes to core
This is a preliminary task to be able to use tus.io in FRS REST API.

In order to not submit a fat review, I propose to refactor code now so
that most of the Upload classes that are used for file upload with tus
(and are not specifically tied to docman) are moved to Tus\Upload
namespace.

No functional changes.

Part of story #12813: upload files with FRS REST API

Change-Id: I3d823be58441cbd8b41b4854cd1b41d2d129f79f