Refactoring: move at a dedicated place the global properties used to create a modal.
This is a part of the  story #12453 Update document modal for existing items

Some modal properties are  used for the creation and the update of an item.
Before the patch the footer, header and the feedback of a modal were in the NewItem folder.
Since their properties are also used to update items, keep their components in the NewItem folder was not relevant anymore.

Their is no functionnal changes expected here

Change-Id: I1ce0c4c9dbbdfc8388a14ceabf52c1c56fc250f7