Display relevant project icon for restricted platform.
Given the platform allows restricted users,

And the variable $feature_flag_project_without_restricted is set to 1,

And there is a project in one of the following visibility:
- public
- public incl. restricted
- private
- private incl. restricted

Then the lock icon is reflecting the visibility in the following places:
☑ sidebar (FP/BP)
☑ navbar (FP/BP)
☑ navbar » projects dropdown (BP)
☑ project dashboard
☑ project » admin » details » access (when platform forbid project admin
to set visibility)
☑ siteadmin » pending projects
☑ siteadmin » projects list
☑ siteadmin » projects list » project

No changes if the feature flag is off.

This introduces icons font in TLP for our custom icons. The set of icons
is displayed in TLP documentation » visual assets » iconography.

Part of story #12554: have private projects without restricted members

Change-Id: I2cbd5a5ba1bcdea8e99ca9b203be633e278c8021