request #13117 User can update an artifact keeping the same integer value
Given a tracker with only an integer field and given an artifact with a value
for this field. A user can open the artifact for edition, open the field and
then submit the artifact without changing the value.

A new changeset is created with the same value, instead of having
a message saying that there are no changes.

This check must works for both Web UI and REST API.

Change-Id: I3d02ec44af69de77dfca1e69eee41d3f8c3eca20