request #13119 User can update an artifact keeping the same float value
Given a tracker with only a float field and given an artifact with a value
for this field. A user can open the artifact for edition, open the field and
then submit the artifact without changing the value.

A new changeset is created with the same value,
instead of having a message saying that there are no changes.

This check must work for both Web UI and REST API.

Change-Id: I5ac560aa9a13120fabdc754eba6567f11a6857f6