request #13127 User can update an artifact keeping the same permission on artifact value
Given a tracker with only a permission on artifact field and given an artifact
with a value for this field. A user can open the artifact for edition,
open the field and then submit the artifact without changing the value.

A new changeset is created with the same value,
instead of having a message saying that there are no changes.

This check must work for both Web UI and REST API.

Additional note: the object artifact was modified without any good
reasons in UpdateArtifact. This update (which is broken because the
searched key in the request never exists) has been removed.

Change-Id: I4a2dd812a4f8a4e05ee4e37229013f1301bc0423