Forbid access to a private project to restricted member
Given the platform allows restricted users
And the variable $feature_flag_project_without_restricted is set to 1,

When the project has visibility private
And user is restricted and still member of this project,

Then user get a 404 error when it tries to access to the project,
And project is not listed in MyProjects widget,
And git operations are rejected.

No changes if the project has visibility private incl. restricted, or if
the feature flag is off.

Please note that SVN, CVS, and REST API will be done in dedicated
changesets.

Part of story #12554: have private projects without restricted members

Change-Id: I1840836985aef716912c565a83920a4157879292