Service knows how to label itself
No need to reinvent the wheel when we instantiate the services for a
project: the logic to translate the label/description of the service is
already implemented in the Service class.

This contribution is needed to avoid to display
"plugin_testmanagement:service_lbl_key" in the project sidebar for TTM
plugin.

*Note:* This is needed for https://gerrit.tuleap.net/c/plugin-testmgmt/+/14459

Part of request #13105: Convert TTM plugin to gettext

Change-Id: Ib374ffa7e3a87f8f35ffd3ff541ec20200d00f43