Display new icons in MyProjects widget.
Given the platform allows restricted users
And the variable $feature_flag_project_without_restricted is set to 1,

And you have projects with the following visibility:
- Private
- Private incl. restricted
- Public
- Public incl. restricted

When the MyProjects widget is added to the personal dashboard,

Then you should see new icons, and the legend is accurate.

No functional changes when the feature flag is off.

Part of story #12554: have private projects without restricted members

Change-Id: I1b03ee41b4a140672f5393bcf4e06d15857ada42