Do not expose projects in navbar, project creation, and rest
Given the platform allows restricted users
And the variable $feature_flag_project_without_restricted is set to 1,

When the project has visibility private
And user is restricted and still member and admin of this project,

Then the project does not appear in:
- projects dropdown (in navbar) (both FP & BP)
- project creation page
- api rest: GET /projects

No changes if the project has visibility private incl. restricted, or if
the feature flag is off.

Part of story #12554: have private projects without restricted members

Change-Id: Icdc42e6e18165f527c5c947987735f227ca3cd68