Prevent edit in place in Cardwall for read-only fields
Part of story #12189 prevent field modification according to state

How to test:
Forge your DB:
- Get a transition id (for example, from "To Do" to "On going") and
insert a new row in
plugin_tracker_workflow_transition_postactions_read_only
- Get field ids for remaining_effort and assigned_to fields
- Get the id you just inserted, get the field ids above and insert new
rows in
plugin_tracker_workflow_transition_postactions_read_only_fields

- Make sure Remaining Effort and Assigned to are selected in Card
semantics

- Browse to a cardwall containing artifacts from your configured tracker
- Remaining effort and Assigned to should not be editable for artifacts
in "On going" status. It should be editable for artifacts in other
status (or when the workflow is disabled).

Note to reviewer:
You can still move cards to another column. The "Edit-in-place" stays
read-only or editable and does not adapt to the new column. PO says this
behaviour is acceptable until the overhaul of Cardwall.

Change-Id: I8523449c7a1bcb25cf9f9dc9e2b7f1f50d5c936f