Extract docman file download from the rest of the docman plugin "routing"
This will allow us to re-use the same approach than the one
used when uploading file from the REST API but for downloading file.

In order to achieve that, the way links are generated within the document
manager plugin has been regrouped in only one place.

No functionnal changes.

Part of request #13133: A REST API user can not download a document manager file

Change-Id: I8d11a4126ecb450b20e3f68550f2fa4c61657207