Remove unused substitution string in OIDC
In acedcbefc4d9efff18479a6af4f7d2bc95ae14bb, a substitution
string should have been removed, because it is not used.

This was not the case before because BaseLanguage::getText silently
remove unused substitution strings.

To test, connect with an OIDC provider, once you are redirected on the
Tuleap interface the button to create a new account should not have
gibberish %1$s.

We should update the tool tab2gettext to prevent to end up in such
situation.

Part of request #13101: Convert OIDC plugin to gettext

Change-Id: I6008617d3d919125a479560ff2e98516d0be54d4