HTTP requests can be processed through a PSR-15 [0] middleware pipeline
This contribution introduce a way to process a PSR-7 [1] server request
using only PSR-15 middlewares. This approach helps the developer to separate
the different layers of the request processing and as a consequence it makes
the request processing easier to test.
This approach is also the one taken in the modern PHP ecosystem so it should
not surprise new developers deep diving into the codebase.

For now, this contribution makes the usage of a PSR-15 middleware pipeline opt-in.
The routes available to upload files in the FRS and document manager has been converted
to it as an example and because they were already using PSR-7 and PSR-15 interfaces.

No functionnal change is expected.

Part of request #13133: A REST API user can not download a document manager file

[0] https://www.php-fig.org/psr/psr-15/
[1] https://www.php-fig.org/psr/psr-7/

Change-Id: I28e173b3ca99331d674d1ae8972e616d9cc029aa