request #11513 Warning when we try to remove a graph
It seems to happen only for the charts bound to empty trackers. When we
try to delete a chart, we get a warning "Illegal string offset 'rank' in
/usr/share/tuleap/plugins/graphontrackersv5/include/data-access/GraphOnTrackersV5_ChartFactory.class.php
on line 106"

This is caused by the fact that the chartFactory tries to sort the
charts retrieved from the session.

How to reproduce:
- Create a brand new tracker based on the bug tracker template
- Go to the "charts" tab
- Try to delete one of the empty charts
--> you get the warning
--> The display is broken

How to test:
--> You can delete / edit / cancel edition without any error

Change-Id: I2f61c2171e56630ca76552b266c25fd2be64c12b