Project admin user autocompleters do not show restricted users when the project is private without restricted
This contribution only affects the autocompleters, nothing else changes.
Note that when a project is is ins the "private without restricted" mode, LDAP
users are not proposed in the autocompleters anymore since we do not know
if the user is going to be restricted or not once the account created.

To test, you will need to enable the "private without restricted" mode by setting
the global parameter feature_flag_project_without_restricted to 1.

This is part of story #13237: Prevent restricted to be added to private projects

Change-Id: I9ae8a2658499973d7a1dfcad24455d368c1c7e65