Have dedicated namespaces for button/dropdown
part of story #13217 Display embedded content

The aim of this commit is to group button/dropdown in dedicated namespace
Actually they can be in modals/quicklook/dropdown...
It makes it hard to add a new button elsewhere (for embedded display, for example),
because this components are too much linked to the place they are actually used.

no functionnal changes.

Change-Id: I888425ae0bbcad5ec5f5a715a8b3e2db7e0f981a