request #13258: Autoloaders generated by Composer for the docman and test management plugins conflict
Both plugins want to load an include/constants.php file. In order to do that, Composer
generates a file identifier from the package name and the path given in the configuration [0].
The file will then only be loaded if the file identifier has not been previously encountered
during the loading process.

In the case of the docman and test management plugins, the path of the file to load is identical
and no package name is defined in the configurations. As a result, Composer will attribute the
name "__root__" to both packages and the two files get the same identifier. This cause issues
when the two files must be loaded simultaneously because only one will be loaded.

To solve that, a package name has been attributed to the packages.

[0] https://github.com/composer/composer/blob/1.8.5/src/Composer/Autoload/AutoloadGenerator.php#L903

Change-Id: I2d105535f4ba06d69fca639db2b76fb23d37160b