Refactor the data build initilization of the rest tests of the Docman
This is a part of story #12638 Add always there metadata date in new/update item modal

This commit put in a dedicated object the data initilization of the docman test rest

Why?
Because in the next commit we have to create a second docman project.
We can't update the database on the fly to set the value of the usage of a metadata.
Then new project will be used to test metadata when metadata is activated.
The actual project will test the all the use case where the metadata are not activated

Change-Id: Ic5fe103aa107940de96181ee229ebcb3ba4b7506