Sync of the project members to an LDAP directory group is aware of the restricted status
In a private project not including restricted users, the restricted users
are not added to the project members during the synchronisation.
The update modal also take this information into account and displays accurate
information.

To test, you will need to enable the "private without restricted" mode by setting
the global parameter feature_flag_project_without_restricted to 1.

This is part of story #13237: Prevent restricted to be added to private projects

Change-Id: Ifa978d1eef6320b706b4b0f7ecc52c48f8b6fcf8