Restricted user cannot be added to project in the 'private without restricted mode'
Both the project members and the user groups are impacted by this change.

To test using the web UI, you can select a non restricted user in the autocompleter,
mark this user as restricted and then try to add the previously selected user.

To test, you will also need to enable the "private without restricted" mode by setting
the global parameter feature_flag_project_without_restricted to 1.

This is part of story #13237: Prevent restricted to be added to private projects

Change-Id: I202e9a2878de13a074154f64322dc6db31a978e4