request #13270: The assigned to me flag in notification subject does not work when the user is also part of a ugroup in a global notification
To reproduce the issue:
 1. Enable the assigned to me feature in a tracker
 2. Choose a user group in the global tracker notification
 3. Assign an artifact to someone part of this user group

The issue is caused by the fact that object reference are compared to
determine if the user is assigned or not to the artifact.
It works when the user is not in the global notification because the same cache
will be used so there is only object representing the user. However,
when the recipients are retrieved from the global notifications the users are
retrieved from the email of the user. In this case we end up with two different
objects representing the same user, the reference comparison fails and the user is
not detected has assigned to the artifact.
For more information check out how UserManager::getUserByUserName/getUserById works
versus UserManager::getUserByEmail.

Change-Id: I682e6847257e09af574f30959f718361d84aedcb