Remove restricted users from project to be private without restricted
The project members are cleaned but all the members of ugroups of this
project are also purged from the restricted users.
A confirmation is requested by a modal when at least one user is going
to be removed by the roject visibility change.

The purge of the restricted users is done in the system event
PROJECT_IS_PRIVATE as it might take a while on large projects.

To test, you will need to enable the "private without restricted" mode by setting
the global parameter feature_flag_project_without_restricted to 1.

This is part of story #13237: Prevent restricted to be added to private projects

Change-Id: I80f3232db0207cac6bf7358eeb492ed68dcefca2